requirements.txt
python-registry==1.0.1
